New templates for different end points are added
Html template for the end points added. 403 template is serve when error occurs, _menu template consist of menu bar, admin page template is a kind of resource which an admin can only access, client info page template will be use to show client id and client secret to user after successfull client registration. userInfo page is a kind of resource page which can be access by admin or a user, welcome page is home page for the Intermine Dashboard.